music_backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
opencv-python>=4.8.0
textblob>=0.17.1
requests>=2.31.0
django-cors-headers>=4.2.0
django-environ>=0.10.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=20.1.0 